[drape] element size can be 255 bytes at maximum

+++ drape/data_buffer.cpp
@@ -1,6 +1,6 @@
 #include "data_buffer.hpp"
 
-DataBuffer::DataBuffer(uint16_t elementSize, uint16_t capacity)
+DataBuffer::DataBuffer(uint8_t elementSize, uint16_t capacity)
   : GLBuffer(GLBuffer::ElementBuffer, elementSize, capacity)
 {
 }
